+++ lib/fcis_core.dart
@@ -5,6 +5,7 @@ export 'src/action_source.dart';
 export 'src/effect_runner.dart';
 export 'src/fcis_composer.dart';
 export 'src/fcis_core.dart';
+export 'src/fcis_lifecycle.dart';
 export 'src/shell.dart';
 export 'src/state_holder.dart';
 export 'src/updater.dart';